+++ src/Device/sdfuOperations.ts
@@ -14,7 +14,11 @@ import { McuState } from '../../nrfutil/device/setMcuState';
 import logger from '../logging';
 import { AppThunk, RootState } from '../store';
 import { getAppFile } from '../utils/appDirs';
-import { clearWaitForDevice, setWaitForDevice } from './deviceAutoSelectSlice';
+import {
+    clearWaitForDevice,
+    setWaitForDevice,
+    WaitForDeviceWhen,
+} from './deviceAutoSelectSlice';
 import { DeviceSetup, DfuEntry } from './deviceSetup';
 import { openDeviceSetupDialog } from './deviceSetupSlice';
 import { Device } from './deviceSlice';
@@ -147,16 +151,20 @@ const switchToDeviceMode =
         device: Device,
         mcuState: McuState,
         onSuccess: (device: Device) => void,
-        onFail: (reason?: unknown) => void
+        onFail: (reason?: unknown) => void,
+        autoReconnectWhen?: WaitForDeviceWhen
     ): AppThunk =>
     dispatch => {
+        if (autoReconnectWhen === undefined) {
+            autoReconnectWhen =
+                mcuState === 'Application'
+                    ? 'applicationMode'
+                    : 'dfuBootLoaderMode';
+        }
         dispatch(
             setWaitForDevice({
                 timeout: 10000,
-                when:
-                    mcuState === 'Application'
-                        ? 'applicationMode'
-                        : 'dfuBootLoaderMode',
+                when: autoReconnectWhen,
                 once: true,
                 onSuccess,
                 onFail,
@@ -199,7 +207,8 @@ export const switchToApplicationMode =
     (
         device: Device,
         onSuccess: (device: Device) => void,
-        onFail: (reason?: unknown) => void
+        onFail: (reason?: unknown) => void,
+        autoReconnectWhen?: WaitForDeviceWhen
     ): AppThunk =>
     dispatch => {
         if (isDeviceInDFUBootloader(device)) {
@@ -216,7 +225,8 @@ export const switchToApplicationMode =
                             );
                         else onSuccess(d);
                     },
-                    onFail
+                    onFail,
+                    autoReconnectWhen
                 )
             );
         } else {
@@ -408,7 +418,8 @@ const programInDFUBootloader =
         dfu: DfuEntry,
         onProgress: (progress: number, message?: string) => void,
         onSuccess: (device: Device) => void,
-        onFail: (reason?: unknown) => void
+        onFail: (reason?: unknown) => void,
+        autoReconnectAfterProgrammingWhen: WaitForDeviceWhen = 'applicationMode'
     ): AppThunk<RootState, Promise<void>> =>
     async dispatch => {
         logger.debug(`${device.serialNumber} on is now in DFU-Bootloader...`);
@@ -494,7 +505,7 @@ const programInDFUBootloader =
                 dispatch(
                     setWaitForDevice({
                         timeout: DEFAULT_DEVICE_WAIT_TIME,
-                        when: 'applicationMode',
+                        when: autoReconnectAfterProgrammingWhen,
                         once: true,
                         onSuccess,
                         onFail,
@@ -518,7 +529,8 @@ const programDeviceWithFw =
     (
         device: Device,
         selectedFw: DfuEntry,
-        onProgress: (progress: number, message?: string) => void
+        onProgress: (progress: number, message?: string) => void,
+        autoReconnectAfterProgrammingWhen: WaitForDeviceWhen = 'applicationMode'
     ): AppThunk<RootState, Promise<Device>> =>
     dispatch =>
         new Promise<Device>((resolve, reject) => {
@@ -529,7 +541,8 @@ const programDeviceWithFw =
                         selectedFw,
                         onProgress,
                         resolve,
-                        reject
+                        reject,
+                        autoReconnectAfterProgrammingWhen
                     )
                 );
                 logger.debug('DFU finished: ', d);
@@ -542,7 +555,8 @@ const programDeviceWithFw =
 
 export const sdfuDeviceSetup = (
     dfuFirmware: DfuEntry[],
-    needSerialport = false
+    needSerialport = false,
+    autoReconnectAfterProgrammingWhen: WaitForDeviceWhen = 'applicationMode'
 ): DeviceSetup => ({
     supportsProgrammingMode: (device, deviceInfo) =>
         ((!!deviceInfo?.dfuTriggerVersion &&
@@ -555,7 +569,12 @@ export const sdfuDeviceSetup = (
             description: firmwareOption.description,
             programDevice: onProgress => dispatch =>
                 dispatch(
-                    programDeviceWithFw(device, firmwareOption, onProgress)
+                    programDeviceWithFw(
+                        device,
+                        firmwareOption,
+                        onProgress,
+                        autoReconnectAfterProgrammingWhen
+                    )
                 ),
         })),
     isExpectedFirmware: (device, deviceInfo) => () => {
@@ -580,7 +599,14 @@ export const sdfuDeviceSetup = (
     },
     tryToSwitchToApplicationMode: device => dispatch =>
         new Promise<Device>((resolve, reject) => {
-            dispatch(switchToApplicationMode(device, resolve, reject));
+            dispatch(
+                switchToApplicationMode(
+                    device,
+                    resolve,
+                    reject,
+                    autoReconnectAfterProgrammingWhen
+                )
+            );
         }),
 });